chore(error): change error display and cleanup
Use of errno and messages should be the same as bash.
Also updating the return.code each time the function is called.

Did some cleanup on a lot of different parts of the code, especially
notes and errors related stuff.

Add initialization on t_var to avoid stupid segfaults cases.
Add `safe_free` function to avoid double free and errors.